Fix cargo clippy feedback.

+++ chirpstack/src/storage/device_profile.rs
@@ -109,7 +109,7 @@ impl DeviceProfile {
 
         if let Some(class_b_params) = &self.class_b_params {
             ds.class_b_ping_slot_dr = class_b_params.ping_slot_dr as u32;
-            ds.class_b_ping_slot_freq = class_b_params.ping_slot_freq as u32;
+            ds.class_b_ping_slot_freq = class_b_params.ping_slot_freq;
             ds.class_b_ping_slot_nb = 1 << class_b_params.ping_slot_nb_k as u32;
         }
 
@@ -133,7 +133,7 @@ impl DeviceProfile {
                 ds.rx1_delay = abp_params.rx1_delay as u32;
                 ds.rx1_dr_offset = abp_params.rx1_dr_offset as u32;
                 ds.rx2_dr = abp_params.rx2_dr as u32;
-                ds.rx2_frequency = abp_params.rx2_freq as u32;
+                ds.rx2_frequency = abp_params.rx2_freq;
             }
         }
     }
